add AbstractProduct and refactor Traits name
