+++ src/Domain/Entity/AbstractProduct.php
@@ -1,0 +1,73 @@
+<?php
+
+namespace App\Domain\Entity;
+
+/**
+ * Class AbstractProduct
+ * @author ereshkidal
+ */
+abstract class AbstractProduct
+{
+    /**
+     * @var string
+     */
+    protected $description;
+
+    /**
+     * @var int
+     */
+    protected $price;
+
+    /**
+     * @var int
+     */
+    protected $stock;
+
+    /**
+     * @return null|string
+     */
+    public function getDescription(): ?string
+    {
+        return $this->description;
+    }
+
+    /**
+     * @param string $description
+     */
+    public function setDescription(string $description): void
+    {
+        $this->description = $description;
+    }
+
+    /**
+     * @return null|int
+     */
+    public function getPrice(): ?int
+    {
+        return $this->price;
+    }
+
+    /**
+     * @param int $price
+     */
+    public function setPrice(int $price): void
+    {
+        $this->price = $price;
+    }
+
+    /**
+     * @return null|int
+     */
+    public function getStock(): ?int
+    {
+        return $this->stock;
+    }
+
+    /**
+     * @param int $stock
+     */
+    public function setStock(int $stock): void
+    {
+        $this->stock = $stock;
+    }
+}
