+++ src/Domain/Entity/Phone.php
@@ -3,30 +3,17 @@
 namespace App\Domain\Entity;
 
 use App\Domain\Entity\Interfaces\EntityInterface;
-use Ramsey\Uuid\Uuid;
-use Ramsey\Uuid\UuidInterface;
+use App\Domain\Entity\Traits\TimestampableTrait;
+use App\Domain\Entity\Traits\UuidTrait;
 
 /**
- * @todo make Phone extends Product and use Traits
  * Class Phone
  * @author ereshkidal
  */
-class Phone implements EntityInterface
+final class Phone extends AbstractProduct implements EntityInterface
 {
-    /**
-     * @var UuidInterface
-     */
-    private $id;
-
-    /**
-     * @var int
-     */
-    private $createdAt;
-
-    /**
-     * @var int
-     */
-    private $updatedAt;
+    use UuidTrait;
+    use TimestampableTrait;
 
     /**
      * @var Manufacturer
@@ -39,101 +26,13 @@ class Phone implements EntityInterface
     private $model;
 
     /**
-     * @var string
-     */
-    private $description;
-
-    /**
-     * @var int
-     */
-    private $price;
-
-    /**
-     * @var int
-     */
-    private $stock;
-
-    /**
      * Phone constructor.
      * @throws \Exception
      */
     public function __construct()
     {
-        $this->id = Uuid::uuid4();
-        $this->createdAt = time();
-        $this->updatedAt = time();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getId(): UuidInterface
-    {
-        return $this->id;
-    }
-
-    /**
-     * @return null|string
-     */
-    public function getModel(): ?string
-    {
-        return $this->model;
-    }
-
-    /**
-     * @param string $model
-     */
-    public function setModel(string $model): void
-    {
-        $this->model = $model;
-    }
-
-    /**
-     * @return null|string
-     */
-    public function getDescription(): ?string
-    {
-        return $this->description;
-    }
-
-    /**
-     * @param string $description
-     */
-    public function setDescription(string $description): void
-    {
-        $this->description = $description;
-    }
-
-    /**
-     * @return null|int
-     */
-    public function getPrice(): ?int
-    {
-        return $this->price;
-    }
-
-    /**
-     * @param int $price
-     */
-    public function setPrice(int $price): void
-    {
-        $this->price = $price;
-    }
-
-    /**
-     * @return null|int
-     */
-    public function getStock(): ?int
-    {
-        return $this->stock;
-    }
-
-    /**
-     * @param int $stock
-     */
-    public function setStock(int $stock): void
-    {
-        $this->stock = $stock;
+        $this->initUuid();
+        $this->initTimestampable();
     }
 
     /**
@@ -150,5 +49,21 @@ class Phone implements EntityInterface
     public function setManufacturer(Manufacturer $manufacturer): void
     {
         $this->manufacturer = $manufacturer;
+    }
+
+    /**
+     * @return null|string
+     */
+    public function getModel(): ?string
+    {
+        return $this->model;
+    }
+
+    /**
+     * @param string $model
+     */
+    public function setModel(string $model): void
+    {
+        $this->model = $model;
     }
 }
